move sock_path
