+++ rust/headless-client/src/ipc.rs
@@ -1,6 +1,6 @@
 #[cfg(target_os = "linux")]
 #[path = "ipc/linux.rs"]
-mod platform;
+pub mod platform;
 
 #[cfg(target_os = "windows")]
 #[path = "ipc/windows.rs"]
