+++ rust/headless-client/src/ipc/linux.rs
@@ -1,6 +1,6 @@
 use super::ServiceId;
 use anyhow::{Context as _, Result};
-use std::os::unix::fs::PermissionsExt;
+use std::{os::unix::fs::PermissionsExt, path::PathBuf};
 use tokio::net::{UnixListener, UnixStream};
 
 pub(crate) struct Server {
@@ -13,7 +13,7 @@ pub(crate) type Stream = UnixStream;
 impl Server {
     /// Platform-specific setup
     pub(crate) async fn new(id: ServiceId) -> Result<Self> {
-        let sock_path = crate::platform::sock_path(id);
+        let sock_path = sock_path(id);
         // Remove the socket if a previous run left it there
         tokio::fs::remove_file(&sock_path).await.ok();
         // Create the dir if possible, needed for test paths under `/run/user`
@@ -40,5 +40,23 @@ impl Server {
             "Accepted an IPC connection"
         );
         Ok(stream)
+    }
+}
+
+/// The path for our Unix Domain Socket
+///
+/// Docker keeps theirs in `/run` and also appears to use filesystem permissions
+/// for security, so we're following their lead. `/run` and `/var/run` are symlinked
+/// on some systems, `/run` should be the newer version.
+///
+/// Also systemd can create this dir with the `RuntimeDir=` directive which is nice.
+pub fn sock_path(id: ServiceId) -> PathBuf {
+    match id {
+        ServiceId::Prod => PathBuf::from("/run")
+            .join(connlib_shared::BUNDLE_ID)
+            .join("ipc.sock"),
+        ServiceId::Test(id) => crate::known_dirs::runtime()
+            .expect("`runtime_dir` should always be computable")
+            .join(format!("ipc_test_{id}.sock")),
     }
 }
